Add files via upload
Used to convert between the fault current and voltage in phase and sequence domains under the symmetrical components model for asymmetric faults. For easy vector addition, do Sum(Vtn) or Sum(Vabc), for the majority of the cases so far, these were approximately be equal to zero in the course.